requirements.txt
google-generativeai==0.3.2
pillow==10.2.0 